move heatmap to plant_layer
