+++ backend/src/model/entity/plant_layer.rs
@@ -1,6 +1,11 @@
 //! Contains the database implementation of the plant layer.
 
-use diesel::{debug_query, pg::Pg, CombineDsl, ExpressionMethods, QueryDsl, QueryResult};
+use diesel::{
+    debug_query,
+    pg::Pg,
+    sql_types::{Double, Integer},
+    CombineDsl, ExpressionMethods, QueryDsl, QueryResult, QueryableByName,
+};
 use diesel_async::{AsyncPgConnection, RunQueryDsl};
 use log::debug;
 
@@ -11,6 +16,41 @@ use crate::{
     },
     schema::relations,
 };
+
+#[derive(Debug, Clone, QueryableByName)]
+pub struct HeatMapElement {
+    #[diesel(sql_type = Double)]
+    pub score: f64,
+    #[diesel(sql_type = Integer)]
+    pub x: i32,
+    #[diesel(sql_type = Integer)]
+    pub y: i32,
+}
+
+pub async fn heatmap(
+    map_id: i32,
+    plant_id: i32,
+    conn: &mut AsyncPgConnection,
+) -> QueryResult<Vec<Vec<f64>>> {
+    // TODO: Compute from the maps geometry
+    let num_rows = 10; // TODO: Calculate number of rows
+    let num_cols = 10; // TODO: Calculate number of columns
+
+    let query = diesel::sql_query("SELECT * FROM calculate_score($1, $2, $3, $4)")
+        .bind::<Integer, _>(map_id)
+        .bind::<Integer, _>(plant_id)
+        .bind::<Integer, _>(num_rows)
+        .bind::<Integer, _>(num_cols);
+
+    let result = query.load::<HeatMapElement>(conn).await?;
+
+    // TODO: figure out how to handle actual values (return matrix to frontend, create image, return matrix as binary?)
+    let mut heatmap = vec![vec![0.0; num_cols as usize]; num_rows as usize];
+    for HeatMapElement { score, x, y } in result {
+        heatmap[x as usize][y as usize] = score;
+    }
+    Ok(heatmap)
+}
 
 /// Get all relations of a certain plant.
 ///
